:recycle: refactor `websockets` module
