+++ fastapi_async_langchain/websockets/__init__.py
@@ -1,10 +1,3 @@
-from .base import BaseLangchainWebsocketConnection
-from .llm import LLMChainWebsocketConnection
-from .retrieval_qa import RetrievalQAWebsocketConnection
+from .base import WebsocketConnection
 
-__all__ = [
-    "BaseLangchainWebsocketConnection",
-    "LLMChainWebsocketConnection",
-    "RetrievalQAWebsocketConnection",
-    "ConversationalRetrievalWebsocketConnection",
-]
+__all__ = ["WebsocketConnection"]
